Added OracleLinux
Due to using OracleLinux at work,  the facter call '$::operatingsystem' returns OracleLinux which isnt seen as a RedHat variant in this manifest.  I added OracleLinux to lines 89 and the if statement  at line 93 to allow OracleLinux to call the yum repo and install the agent correctly across our server fleet.