Enable upgrading of Windows agent versions
The Windows install currently only realistically allows the installation of whatever version the upstream newrelic-infra.msi is, and doesn't allow for the possibility of upgrading the newrelic agent once it is present on a server.

This patch only uses the file name "newrelic-infra.msi" in the case that the version is set to be "installed" or "latest". For other versions, it tries to download and install a matching newrelic-infra.${ensure}.msi file and install that.

While this doesn't enable the real use of "latest" as a version, it does enable a way to pick a specific version and so ensure that upgrades of agents become possible.