Bugfix to specified agent download on Windows
This fixes a bug in which version of the NewRelic agent gets installed on Windows.

When a specific version of the agent is configured to be installed, the previous download_url specifically pointed at `newrelic-infra.msi` and *saved* it as a file with the name of the specific version. In most cases the specific version would be the current version, i.e. whatever `newrelic-infra.msi` was then anyway, so this bug was hidden.

If however a newer agent version had been released upstream (e.g. 1.24.2) than what was configured to be installed (e.g. 1.24.1), and a fresh server was built and ran puppet, it would download newrelic-infra.msi (i.e. 1.24.2), but save it as newrelic-infra-1.24.1.msi, and then install that misnamed package from disk.